% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{Gapminder Dataset (1952 -- 2012)}
\format{An object of class \code{data.frame} with 3577 rows and 55 columns.}
\source{
\url{http://www.gapminder.org/data/}
}
\usage{
gapminder
}
\description{
A dataset containing a variety of indiators from the Gapminder website,
recorded every four years from 1952 to 2012.
Full details can be found on the website.
}
\keyword{datasets}

