% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gapminder_2008}
\alias{gapminder_2008}
\title{Gapminder Dataset (2008)}
\format{An object of class \code{data.frame} with 225 rows and 27 columns.}
\source{
\url{http://www.gapminder.org/data/}
}
\usage{
gapminder_2008
}
\description{
A dataset containing a variety of indiators from the Gapminder website,
limited to the year 2008. Full details can be found on the website.
}
\keyword{datasets}
