% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes2009_2012}
\alias{nhanes2009_2012}
\title{NHANES Dataset (2009 -- 2012)}
\format{An object of class \code{data.frame} with 10000 rows and 75 columns.}
\source{
\url{http://www.cdc.gov/nchs/nhanes/nhanes_questionnaires.htm}
}
\usage{
nhanes2009_2012
}
\description{
A dataset adapted from the publicly available NHANES surveys from
2009--2012
This is survey data collected by the US National Center for Health Statistics
(NCHS) which has conducted a series of health and nutrition surveys since the
early 1960's. Since 1999 approximately 5,000 individuals of all ages are
interviewed in their homes every year and complete the health examination
component of the survey. The health examination is conducted in a mobile
examination centre (MEC).
}
\keyword{datasets}

