% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes_1000}
\alias{nhanes_1000}
\title{NHANES Dataset (First 1000 rows)}
\format{An object of class \code{data.frame} with 1000 rows and 41 columns.}
\source{
\url{http://www.cdc.gov/nchs/nhanes/nhanes_questionnaires.htm}
}
\usage{
nhanes_1000
}
\description{
The first 1000 rows of the NHANES data set.
}
\keyword{datasets}
