% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes_1000_50}
\alias{nhanes_1000_50}
\title{NHANES Dataset (First 50 rows)}
\format{An object of class \code{data.frame} with 50 rows and 41 columns.}
\source{
\url{http://www.cdc.gov/nchs/nhanes/nhanes_questionnaires.htm}
}
\usage{
nhanes_1000_50
}
\description{
The first 50 rows of the NHANES data set.
}
\keyword{datasets}

