% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes_2000}
\alias{nhanes_2000}
\title{NHANES Dataset (First 2000 rows)}
\format{An object of class \code{data.frame} with 2000 rows and 74 columns.}
\source{
\url{http://www.cdc.gov/nchs/nhanes/nhanes_questionnaires.htm}
}
\usage{
nhanes_2000
}
\description{
The first 2000 rows of the NHANES data set.
}
\keyword{datasets}

