% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{olympics100m}
\alias{olympics100m}
\title{Olympics 100m}
\format{An object of class \code{data.frame} with 49 rows and 5 columns.}
\usage{
olympics100m
}
\description{
Data from the 100m olympics event.
}
\keyword{datasets}
