% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{week7_CanadaPop}
\alias{week7_CanadaPop}
\title{Canada Population}
\format{An object of class \code{data.frame} with 19 rows and 2 columns.}
\usage{
week7_CanadaPop
}
\description{
This data is based on an experiment where the researcher asked 19
students to estimate (in millions) the population of Canada
(see Utts, J. M. & Heckard, R. F. (2014). Mind on Statistics,
(5th ed.). USA: Cengage Learning.)
}
\keyword{datasets}
