% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{week7_CocaineTrtmtExpt_2Trt_D_P}
\alias{week7_CocaineTrtmtExpt_2Trt_D_P}
\title{Cocaine Treatment Experiment (Desipramine/Placebo)}
\format{An object of class \code{data.frame} with 48 rows and 3 columns.}
\usage{
week7_CocaineTrtmtExpt_2Trt_D_P
}
\description{
Only the Desipramine and Placebo treatments.
}
\details{
From data from Gawin, F. et al (1989)
Archives of General Psychiatry 46(2) (117-121)
found in Statistics Unlocking the Power of Data (2013)
by Lock, R.H. et al.
}
\keyword{datasets}

