% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{week7_LightatNight}
\alias{week7_LightatNight}
\title{Light at Night for Mice}
\format{An object of class \code{data.frame} with 27 rows and 11 columns.}
\source{
\url{https://cran.r-project.org/web/packages/Lock5Data/}
}
\usage{
week7_LightatNight
}
\description{
Data from an experiment with mice having different nighttime
light conditions.
}
\keyword{datasets}
