% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{week7_MindsetMatters}
\alias{week7_MindsetMatters}
\title{Mindset Matters}
\format{An object of class \code{data.frame} with 75 rows and 20 columns.}
\usage{
week7_MindsetMatters
}
\description{
This data is based on an experiment where the researcher was
interested in the placebo effects of mind-set on the body.
84 female work attendants from 7 large hotels were randomly
allocated to one of two groups, Informed and Control.
(Those with missing values have been taken out of our dataset.)
}
\keyword{datasets}

