% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{week7_SleepDeprivationExpt}
\alias{week7_SleepDeprivationExpt}
\title{Light at Night for Mice}
\format{An object of class \code{data.frame} with 21 rows and 2 columns.}
\usage{
week7_SleepDeprivationExpt
}
\description{
Nature Neuroscience 2 (1237-8)
given in Investigating Statistical Concepts Applications and
Methods (2006) by Rossman and Chance
}
\keyword{datasets}

