% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{week8_Actual_VisitorArrivalsQuarterlyFrom1998}
\alias{week8_Actual_VisitorArrivalsQuarterlyFrom1998}
\title{Quarterly Visitor Arrivals from 1998}
\format{An object of class \code{data.frame} with 66 rows and 13 columns.}
\source{
Statistics New Zealand
}
\usage{
week8_Actual_VisitorArrivalsQuarterlyFrom1998
}
\description{
Visitor arrivals by country of residence all travel mode
(Quarterly)
}
\details{
Visitor arrivals are overseas residents arriving in New Zealand
for a stay of less than 12 months.
For arrival series, the country of residence is the country where
a person arriving in New Zealand last lived for 12 months or
more (country of last permanent residence).

The United Kingdom also includes the Channel Islands and the
Isle of Man.

Cruise passengers have only been collected separately since
September 1997.
Cruise passengers are only recorded as arrivals or departures
if they begin or end their cruise in New Zealand.
Otherwise, they are regarded as being in transit,
even if they leave the ship for periods while it is in New Zealand.
}
\keyword{datasets}

