% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{week8_AverageVisitorsMonthly}
\alias{week8_AverageVisitorsMonthly}
\title{Average Monthly Visitor Arrivals}
\format{An object of class \code{data.frame} with 190 rows and 9 columns.}
\source{
Statistics New Zealand
}
\usage{
week8_AverageVisitorsMonthly
}
\description{
Average number of visitors in New Zealand by country of
residence (Monthly)
The United Kingdom also includes the Channel Islands and the
Isle of Man.
The number of visitors in New Zealand is estimated from a
sample of records, using day of arrival and intended length
of stay. It is calculated for each day within the selected
period and averaged.
}
\keyword{datasets}
