% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{week8_AverageVisitorsQuarterly}
\alias{week8_AverageVisitorsQuarterly}
\title{Average Quarterly Visitor Arrivals}
\format{An object of class \code{data.frame} with 63 rows and 9 columns.}
\source{
Statistics New Zealand
}
\usage{
week8_AverageVisitorsQuarterly
}
\description{
Average number of visitors in New Zealand by country of
residence (Quarterly)
}
\details{
The United Kingdom also includes the Channel Islands and the
Isle of Man.
The number of visitors in New Zealand is estimated from a
sample of records, using day of arrival and intended length
of stay. It is calculated for each day within the selected
period and averaged.
}
\keyword{datasets}
